+++ src/components/Expense/Expense.jsx
@@ -13,6 +13,7 @@ import {
 } from "@/redux/actions/expensesActions";
 import { deleteMovement } from "@/redux/actions/movementsActions";
 import SpinnerSmall from "./../../utils/Spinner/SpinnerSmall";
+import Swal from "sweetalert2";
 
 const typeExpenses = [
   "servicios públicos",
@@ -143,7 +144,21 @@ const Expense = () => {
   };
 
   const handleDeleteMovement = (idMovement) => {
-    dispatch(deleteMovement(idMovement));
+    Swal.fire({
+      title: "¿Seguro que desea eliminar este egreso?",
+      showCancelButton: true,
+      confirmButtonText: "Eliminar",
+      cancelButtonText: "Cancelar",
+      reverseButtons: true,
+      customClass: {
+        confirmButton: "my-confirm-button",
+        cancelButton: "my-cancel-button",
+      },
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(deleteMovement(idMovement));
+      }
+    });
   };
   const handleFilterType = () => {
     dispatch(filterExpense(selectType));
@@ -521,7 +536,7 @@ const Expense = () => {
                             className="btn-trash"
                             onClick={() => handleDeleteMovement(expense._id)}
                           >
-                            Eliminar
+                           <svg xmlns="http://www.w3.org/2000/svg" height="13px" viewBox="0 -960 960 960" width="13px" fill="white"><path d="M280-120q-33 0-56.5-23.5T200-200v-520h-40v-80h200v-40h240v40h200v80h-40v520q0 33-23.5 56.5T680-120H280Zm400-600H280v520h400v-520ZM360-280h80v-360h-80v360Zm160 0h80v-360h-80v360ZM280-720v520-520Z"/></svg>
                           </button>
                         </td>
                       )}
