+++ src/components/Income/Income.jsx
@@ -8,6 +8,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { createIncome, getIncomes, filterIncome, cleanFilter } from "@/redux/actions/incomesActions";
 import { deleteMovement } from "@/redux/actions/movementsActions";
 import SpinnerSmall from './../../utils/Spinner/SpinnerSmall';
+import Swal from "sweetalert2";
 
 
 const Income = () => {
@@ -120,7 +121,21 @@ const Income = () => {
   };
 
   const handleDeleteMovement = (idMovement) => {
-    dispatch(deleteMovement(idMovement));
+    Swal.fire({
+      title: "¿Seguro que desea eliminar este ingreso?",
+      showCancelButton: true,
+      confirmButtonText: "Eliminar",
+      cancelButtonText: "Cancelar",
+      reverseButtons: true,
+      customClass: {
+        confirmButton: "my-confirm-button",
+        cancelButton: "my-cancel-button",
+      },
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(deleteMovement(idMovement));
+      }
+    });
   };
   
 
@@ -406,7 +421,7 @@ const Income = () => {
                             className="btn-trash"
                             onClick={() => handleDeleteMovement(income._id)}
                           >
-                            Eliminar
+                           <svg xmlns="http://www.w3.org/2000/svg" height="13px" viewBox="0 -960 960 960" width="13px" fill="white"><path d="M280-120q-33 0-56.5-23.5T200-200v-520h-40v-80h200v-40h240v40h200v80h-40v520q0 33-23.5 56.5T680-120H280Zm400-600H280v520h400v-520ZM360-280h80v-360h-80v360Zm160 0h80v-360h-80v360ZM280-720v520-520Z"/></svg>
                           </button>
                         </td>
                       )}
