+++ src/components/Offices/Offices.jsx
@@ -15,11 +15,12 @@ import {
   updateStockCurrency,
 } from "../../redux/actions/subOfficesActions";
 import "./Offices.css";
+import Swal from "sweetalert2";
 
 const Offices = () => {
   const dispatch = useDispatch();
   const subOffices = useSelector((state) => state.offices.subOffices);
-  console.log(subOffices)
+  console.log(subOffices);
   const currencies = useSelector((state) => state.offices.currencies);
   const updaCurrency = useSelector(
     (state) => state.offices._idupdateCurrencies
@@ -32,7 +33,7 @@ const Offices = () => {
     name: "",
     code: "",
     address: "",
-    phone: ""
+    phone: "",
   });
   const [visibleCurrencies, setVisibleCurrencies] = useState({});
 
@@ -55,7 +56,21 @@ const Offices = () => {
     setModalOpen(false);
   };
   const handleDeleteCurrency = (idCurrency, idOffice) => {
-    dispatch(deleteCurrencySubOffice(idCurrency, idOffice));
+    Swal.fire({
+      title: "¿Seguro que desea eliminar esta moneda?",
+      showCancelButton: true,
+      confirmButtonText: "Eliminar",
+      cancelButtonText: "Cancelar",
+      reverseButtons: true,
+      customClass: {
+        confirmButton: "my-confirm-button",
+        cancelButton: "my-cancel-button",
+      },
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(deleteCurrencySubOffice(idCurrency, idOffice));
+      }
+    });
   };
   const handleSelectChange = (e) => {
     setSelectedCurrencyId(e.target.value);
@@ -285,7 +300,6 @@ const Offices = () => {
                     Ver usuarios de la sucursal <img src={arrowDown} alt="" />
                   </span>
                 </div>*/}
-            
 
                 {/* Seccion monedas */}
                 <div>
@@ -409,7 +423,15 @@ const Offices = () => {
                                     )
                                   }
                                 >
-                                  Eliminar
+                                  <svg
+                                    xmlns="http://www.w3.org/2000/svg"
+                                    height="13px"
+                                    viewBox="0 -960 960 960"
+                                    width="13px"
+                                    fill="white"
+                                  >
+                                    <path d="M280-120q-33 0-56.5-23.5T200-200v-520h-40v-80h200v-40h240v40h200v80h-40v520q0 33-23.5 56.5T680-120H280Zm400-600H280v520h400v-520ZM360-280h80v-360h-80v360Zm160 0h80v-360h-80v360ZM280-720v520-520Z" />
+                                  </svg>
                                 </button>
                               </td>
                             </tr>
